Update development Docker configuration
This is mostly the same as the "testing" configuration, though it uses
the development Dockerfile for the frontend. The frontend PORT is set to
the same port that nginx uses in the "testing" configuration for
consistency.

The subnet is also changed compared to the "testing" configuration, so
that they can both be run on the same machine without issues.