Add dummy SMTP server to development configuration
This avoids potentially accidentally sending mail to users while in the
development configuration, and also makes it easy to see exactly which
emails were sent when (Python's smtpd.DebuggingServer prints all mail
it's asked to send to stdout, which docker-compose makes visible).

This requires the backend's configuration (in /cogs/config/dev.yaml) to
be updated to use the "mailserver" host for sending mail.