Handle authentication in nginx
Previously, the frontend would handle authentication by asking a backend
service called Cookie Thief to tell it the value of the Pagesmith_User
cookie (said cookie is HttpOnly), then passing that value in an
Authorization header to the API. If the cookie didn't exist, the
frontend would set window.location to redirect to the PageSmith login
page. (The API-providing backend could have read the cookie, but for the
sake of decoupling, it didn't.)

This is clearly suboptimal -- if nothing else, the extra network request
involved in interrogating Cookie Thief about the cookie could be avoided
-- so this commit teaches nginx to rewrite the cookie into an
Authorization header before proxying API requests to the backend, as
well as redirecting non-API requests to the login page if the PageSmith
cookie isn't present. (NB that the PageSmith login page makes use of
static assets served by the frontend server, so it's still necessary to
pass those requests through even when the cookie is not present.)

Now, the only case the frontend has to deal with is the case where the
cookie is invalid for whatever reason, in which case the backend will
return a 401 and the frontend can display an error and/or redirect to
/login (currently it does both).

This commit also works around a presumed bug in nginx where explicitly
specifying port 80 in proxy_pass/proxy_redirect directives results in
the proxied requests not being rewritten to change the hostname, meaning
that the client would receive redirects to "frontend" instead of
"sanger.ac.uk": https://trac.nginx.org/nginx/ticket/1804