Banish the Cookie Thief
Since nginx now handles translating the PageSmith cookie into an
Authorization header, the /steal Cookie Thief endpoint is no longer
used. Further, the /munch endpoint essentially serves a static response
consisting of a redirect to /login and a Set-Cookie header that removes
the PageSmith cookie; nginx is perfectly capable of serving a redirect
with a Set-Cookie header on its own, so this commit teaches it to do so.

The Set-Cookie header is not included in development mode, because
PageSmith authentication is not used there.