Avoid sending an invalid Authorization header
If no PageSmith cookie is present, then requests to the frontend are
always redirected to /login, but requests to the backend are not
redirected -- instead, a malformed Authorization header of "Pagesmith "
would be sent, which would cause the authentication middleware to fail
in an ugly way.

Now, nginx only adds an Authorization header if the PageSmith cookie is
actually present, which improves the error message from the API when the
client is not authenticated.