Mount /etc/faketimerc in the development backend container
When a backend with libfaketime installed is used, this can be used to
modify the apparent flow of time inside the container.

NB: using a file (as opposed to a directory) as a volume is error-prone!
If you edit the file while the container is mounted, you must make sure
that your editor does not actually replace the file with a new one,
because that will change the file's inode number, and further updates to
the file will not be synchronised between the host and the container
(until is is restarted, when any container-side changes will be lost).
See moby/moby#15793 for details and workarounds.